+++ src/ts/components/adclickoverlay.ts
@@ -1,6 +1,10 @@
-import {ClickOverlay} from './clickoverlay';
-import {UIInstanceManager} from '../uimanager';
-import { AdEvent, PlayerAPI } from 'bitmovin-player';
+import { ClickOverlay } from './clickoverlay';
+import { UIInstanceManager } from '../uimanager';
+import { Ad, AdEvent, PlayerAPI } from 'bitmovin-player';
+
+interface LocalAd extends Ad {
+  clickThroughUrlOpened?: () => void;
+}
 
 /**
  * A simple click capture overlay for clickThroughUrls of ads.
@@ -10,39 +14,30 @@ export class AdClickOverlay extends ClickOverlay {
   configure(player: PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
-    let clickThroughUrl = <string>null;
-    let clickThroughEnabled = !player.getConfig().advertising
-      || !player.getConfig().advertising.hasOwnProperty('clickThroughEnabled')
-      || (player.getConfig().advertising as any).clickThroughEnabled;
+    let clickThroughCallback: () => void = null;
 
     player.on(player.exports.PlayerEvent.AdStarted, (event: AdEvent) => {
-      clickThroughUrl = null; // TODO event.clickThroughUrl;
-
-      if (clickThroughEnabled) {
-        this.setUrl(clickThroughUrl);
-      } else {
-        // If click-through is disabled, we set the url to null to avoid it open
-        this.setUrl(null);
-      }
+      let ad = event.ad as LocalAd;
+      this.setUrl(ad.clickThroughUrl);
+      clickThroughCallback = ad.clickThroughUrlOpened;
     });
 
     // Clear click-through URL when ad has finished
     let adFinishedHandler = () => {
       this.setUrl(null);
     };
+
     player.on(player.exports.PlayerEvent.AdFinished, adFinishedHandler);
     player.on(player.exports.PlayerEvent.AdSkipped, adFinishedHandler);
     player.on(player.exports.PlayerEvent.AdError, adFinishedHandler);
 
     this.onClick.subscribe(() => {
       // Pause the ad when overlay is clicked
-      player.pause('ui-content-click');
+      player.pause('ui-ad-click-overlay');
 
-      // Notify the player of the clicked ad
-      // TODO add a callback to AdStarted to allow the ads renderer to signal a clickThroughUrl click
-      // player.fireEvent(player.exports.PlayerEvent.AdClicked, {
-      //   clickThroughUrl: clickThroughUrl,
-      // });
+      if (clickThroughCallback) {
+        clickThroughCallback();
+      }
     });
   }
 }